Rename position directory to be singualr
